requirements.txt
notebooklm-py
python-dotenv
